feat: Rename component
