+++ packages/vue-component-library/src/lib-components/SearchMenuItem.vue
